Correct the wire format against real hardware
Verified against a 12-key / 2-knob pad (1189:8842). Four things were wrong
in the version derived from static analysis alone:

- Reports are 65 bytes including the leading 0x03. hidapi only strips that
  first byte when it is zero, so the whole buffer belongs on the wire. The
  interface confirms it: MaxOutputReportSize is 65.
- Reading is per layer, not per key. Read_configuration_clicked calls
  read_Hidkey_Data(3, 0x0F, 3), so the request is 03 FA 0F 03 <layer> and
  the pad answers with a burst of records.
- Media bindings hold a 16-bit HID Consumer usage little-endian across
  bytes 10 and 11 (E9 00 is volume up), not a single invented selector.
- Mouse bindings put 4 in the count field as a payload length.

Knob slots also turned out to be fixed rather than following the last key:
knob n occupies 16 + 3n through 18 + 3n.

Tests now include six reports captured from the device, so the decoder is
pinned to observed behaviour rather than to inference.

+++ Sources/MiniKeyboardKit/Protocol/KeyAction.swift
@@ -50,7 +50,7 @@ public enum KeyAction: Sendable, Hashable, Codable {
     /// One or more chords played in sequence (up to `Wire.maxMacroSteps`).
     case keyboard([Chord])
     /// A consumer-page media control.
-    case media(UInt8)
+    case media(UInt16)
     /// A mouse button bitmask and/or wheel delta.
     case mouse(buttons: UInt8, wheel: Int8)
